refactor: separate lexio divina form component
