+++ pages/contemplation.tsx
@@ -1,18 +1,11 @@
-import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
-import RemoveCircleOutlineIcon from '@mui/icons-material/RemoveCircleOutline';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import IconButton from '@mui/material/IconButton';
-import MenuItem from '@mui/material/MenuItem';
-import Select from '@mui/material/Select';
-import TextField from '@mui/material/TextField';
-import Typography from '@mui/material/Typography';
 import AccessDeniedMessage from 'components/AccessDeniedMessage';
 import GlobalHeader from 'components/GlobalHeader';
 import { GLOBAL_HEADER_HEIGHT } from 'components/GlobalHeader/GlobalHeader';
+import LexioDivinaForm from 'components/LexioDivinaForm/LexioDivinaForm';
 import LoginMessage from 'components/LoginMessage';
 import Meta from 'components/Meta';
-import { Bible, bibleOptions } from 'constants/bible';
+import { Bible } from 'constants/bible';
 import useAuth from 'hooks/context/useAuth';
 import { postLexioDivina, putLexioDivina } from 'libs/axios/apis';
 import { ObjectId } from 'mongodb';
@@ -20,8 +13,7 @@ import type { GetServerSideProps, NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { useSnackbar } from 'notistack';
 import { useState } from 'react';
-import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-import { FormattedMessage } from 'react-intl';
+import { SubmitHandler, useForm } from 'react-hook-form';
 import Database, { CollectionName } from 'server/database';
 import { PRIMARY_SHADOW } from 'theme/shadows';
 import { LexioDivina } from 'types/interfaces';
@@ -110,14 +102,7 @@ const Contemplation: NextPage<ContemplationProps> = ({
   const { enqueueSnackbar } = useSnackbar();
   const { user, logout } = useAuth();
 
-  const {
-    control,
-    register,
-    handleSubmit,
-    getValues,
-    setValue,
-    formState: { errors },
-  } = useForm<FormInput>({
+  const form = useForm<FormInput>({
     defaultValues: defaultValues
       ? defaultValues
       : {
@@ -126,7 +111,6 @@ const Contemplation: NextPage<ContemplationProps> = ({
   });
 
   const [isRequested, setIsRequested] = useState(false);
-  const [isExpanded, setIsExpanded] = useState(!!defaultValues?.endChapter);
 
   const handleCancel = () => {
     router.back();
@@ -182,16 +166,6 @@ const Contemplation: NextPage<ContemplationProps> = ({
     } finally {
       setIsRequested(false);
     }
-  };
-
-  const handleClickPlusButton = () => {
-    setIsExpanded(true);
-  };
-
-  const handleClickMinusButton = () => {
-    setIsExpanded(false);
-    setValue('endChapter', '');
-    setValue('endVerse', '');
   };
 
   // 로그인을 하지 않았을 경우
@@ -262,11 +236,9 @@ const Contemplation: NextPage<ContemplationProps> = ({
 
           {/* right */}
           <Box
-            component="form"
             display="flex"
             flexDirection="column"
             gap={2}
-            onSubmit={handleSubmit(onSubmit)}
             sx={{
               height: '100%',
               overflowY: 'auto',
@@ -277,180 +249,13 @@ const Contemplation: NextPage<ContemplationProps> = ({
               },
             }}
           >
-            <Box
-              sx={{
-                p: 2,
-                display: 'flex',
-                flexDirection: 'column',
-                gap: 1.5,
-              }}
-            >
-              <Box>
-                <TextField
-                  {...register('phrase', {
-                    required: true,
-                  })}
-                  size="small"
-                  fullWidth
-                  multiline
-                  minRows={2}
-                  placeholder="마음에 와닿은 구절"
-                  error={!!errors.phrase}
-                />
-              </Box>
-
-              <Box
-                display="grid"
-                gridTemplateColumns="1fr 1fr 1fr auto"
-                gap={1.5}
-              >
-                <Box>
-                  <Controller
-                    control={control}
-                    name="bible"
-                    render={({ field }) => (
-                      <Select
-                        {...field}
-                        size="small"
-                        defaultValue={Bible.Genesis}
-                        fullWidth
-                      >
-                        {bibleOptions.map((option) => (
-                          <MenuItem key={option.value} value={option.value}>
-                            {option.label}
-                          </MenuItem>
-                        ))}
-                      </Select>
-                    )}
-                  />
-                </Box>
-                <Box>
-                  <TextField
-                    {...register('chapter', {
-                      required: true,
-                      validate: {
-                        moreThanOne: (value) => Number(value) > 0,
-                      },
-                    })}
-                    size="small"
-                    placeholder="장"
-                    fullWidth
-                    type="number"
-                    error={!!errors.chapter}
-                  />
-                </Box>
-                <Box>
-                  <TextField
-                    {...register('verse', {
-                      required: true,
-                      validate: {
-                        moreThanOne: (value) => Number(value) > 0,
-                      },
-                    })}
-                    size="small"
-                    placeholder="절"
-                    fullWidth
-                    type="number"
-                    error={!!errors.verse}
-                  />
-                </Box>
-                <Box ml={-1.5} alignSelf="center">
-                  {isExpanded ? (
-                    <Typography color="text.primary" textAlign="end">
-                      에서
-                    </Typography>
-                  ) : (
-                    <IconButton onClick={handleClickPlusButton}>
-                      <AddCircleOutlineIcon />
-                    </IconButton>
-                  )}
-                </Box>
-
-                {isExpanded ? (
-                  <>
-                    <Box justifySelf="flex-end" alignSelf="center" mr={-1.5}>
-                      <IconButton onClick={handleClickMinusButton}>
-                        <RemoveCircleOutlineIcon />
-                      </IconButton>
-                    </Box>
-                    <Box>
-                      <TextField
-                        {...register('endChapter', {
-                          validate: {
-                            onlyPlus: (value) => !value || Number(value) >= 1,
-                            moreThanStartChapter: (value) =>
-                              !value ||
-                              Number(value) >= Number(getValues('chapter')),
-                            haveBoth: (value) =>
-                              (!!value && !!getValues('endVerse')) ||
-                              (!value && !getValues('endVerse')),
-                          },
-                        })}
-                        size="small"
-                        placeholder="장"
-                        fullWidth
-                        type="number"
-                        error={!!errors.endChapter}
-                      />
-                    </Box>
-                    <Box>
-                      <TextField
-                        {...register('endVerse', {
-                          validate: {
-                            onlyPlus: (value) => !value || Number(value) >= 1,
-                            moreThanStarted: (value) =>
-                              !value ||
-                              Number(getValues('chapter')) <
-                                Number(getValues('endChapter')) ||
-                              Number(value) > Number(getValues('verse')),
-                            haveBoth: (value) =>
-                              (!!value && !!getValues('endChapter')) ||
-                              (!value && !getValues('endChapter')),
-                          },
-                        })}
-                        size="small"
-                        placeholder="절"
-                        fullWidth
-                        type="number"
-                        error={!!errors.endVerse}
-                      />
-                    </Box>
-                    <Box alignSelf="center">
-                      <Typography color="text.primary" textAlign="end">
-                        까지
-                      </Typography>
-                    </Box>
-                  </>
-                ) : null}
-              </Box>
-
-              <Box>
-                <TextField
-                  {...register('content', {
-                    required: true,
-                  })}
-                  size="small"
-                  fullWidth
-                  multiline
-                  minRows={15}
-                  placeholder="구절을 통해 느낀점"
-                  error={!!errors.content}
-                />
-              </Box>
-
-              <Box display="flex" justifyContent="flex-end" gap={1}>
-                <Button variant="outlined" onClick={handleCancel}>
-                  <FormattedMessage id="common.cancel" />
-                </Button>
-                <Button
-                  type="submit"
-                  variant="contained"
-                  disabled={isRequested}
-                >
-                  <FormattedMessage id="common.save" />
-                </Button>
-              </Box>
-            </Box>
+            <LexioDivinaForm
+              form={form}
+              isRequested={isRequested}
+              defaultIsExpanded={!!defaultValues?.endChapter}
+              onCancel={handleCancel}
+              onSubmit={onSubmit}
+            />
           </Box>
         </Box>
       </Box>
